# backend/src/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: PreFix – Maintenance Management API
  version: "1.0.0"
  description: |
    API per la gestione della manutenzione degli edifici comunali.

    **Sicurezza**
    - Tutte le API `/api/v1/*` sono **protette**
    - L’autenticazione reale è gestita da **Proxy + Cookie HTTP-only**
    - L’autorizzazione è gestita tramite **RBAC (Policy Decision Point)**
    - Swagger utilizza **Bearer Token SOLO per documentazione/test manuale**

    Swagger **NON** gestisce correttamente i cookie HTTP-only.
    Per questo motivo:
    - Effettuare il login tramite Postman / Curl / Frontend
    - Inserire manualmente un access token nel pulsante **Authorize** (solo per test)

    **Ruoli**
    - Utente registrato
    - Utente secondario
    - Admin locale
    - Admin centrale
    - Impresa
servers:
  - url: http://localhost:5000
    description: Proxy API Gateway

# Tutte le API sono protette di default
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: Lista utenti
      description: |
        **Ruolo richiesto:** Admin Centrale
        Restituisce la lista di tutti gli utenti
      responses:
        "200":
          description: Lista utenti
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Non autenticato
        "403":
          description: Accesso negato
...
  /api/v1/buildings:
    get:
      tags: [Buildings]
      summary: Edifici associati
      description: |
        **Ruolo richiesto:** Utente registrato
        Restituisce tutti i buildings a cui un utente è associato
      responses:
        "200":
          description: Lista edifici
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Building"
        "401":
          description: Non autenticato
        "403":
          description: Accesso negato


  # =====================
  # EVENTS
  # =====================
  /api/v1/events:
    get:
      tags: [Events]
      summary: Lista eventi
      description: |
        **Ruolo richiesto:** Utente registrato
        Restituisce gli eventi di manutenzione associati agli edifici dell’utente
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [calendar]
          description: |
            Vista degli eventi:
            - `null` (default) → registro completo degli eventi
            - `calendar` → restituisce solo eventi con data pianificata (`scheduledAt != null`)
      responses:
        "200":
          description: Lista eventi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "401":
          description: Non autenticato
...
  /api/v1/interventions:
    get:
      tags: [Interventions]
      summary: Storico interventi
      description: |
        Restituisce lo storico degli interventi associati
        agli edifici dell’utente.

        ### Modalità di utilizzo
        - **Senza parametri:** tutti gli interventi (storico completo)
        - **Con `period`:** interventi filtrati per periodo
        - **Con `assetId`:** interventi relativi a uno specifico asset
        - **Combinabili**
      parameters:
        - name: period
          in: query
          required: false
          description: |
            Filtro temporale sugli interventi.
            Valori ammessi:
            - `month` → ultimo mese
            - `quarter` → ultimo trimestre
            - `year` → ultimo anno
          schema:
            type: string
            enum: [month, quarter, year]

        - name: assetId
          in: query
          required: false
          description: ID dell’asset per filtrare gli interventi
          schema:
            type: string

      responses:
        "200":
          description: Lista interventi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Intervention"

        "400":
          description: Parametri non validi
        "401":
          description: Non autenticato
        "403":
          description: Non autorizzato
...
  /api/v1/requests:
    get:
      tags: [Requests]
      summary: Lista richieste
      description: |
        **Ruolo richiesto:** Admin Centrale
        Restituisce la lista di tutte le richieste presenti nel sistema.
      responses:
        "200":
          description: Lista richieste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Request"
        "401":
          description: Non autenticato
        "403":
          description: Non autorizzato
...
  /api/v1/notifications:
    get:
      tags: [Notifications]
      summary: Notifiche utente
      description: |
        **Ruolo richiesto:** Utente registrato
      parameters:
        - name: not_read
          in: query
          required: false
          schema:
            type: boolean
          description: |
            Se true, restituisce solo le notifiche non lette
      responses:
        "200":
          description: Lista notifiche
        "401":
          description: Non autenticato
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Inserire l'access token JWT.
        Usato SOLO per test su Swagger UI.
...
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
    Building:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
    Event:
      type: object
      properties:
        _id:
          type: string
        assetId:
          type: string
        status:
          type: string
        scheduledAt:
          type: string
          format: date-time
    Intervention:
      type: object
      properties:
        _id:
          type: string

        assetId:
          type: string
          description: Asset associato all’intervento

        type:
          type: string
          enum: [maintenance, inspection, failure, repair]

        outcome:
          type: string
          enum: [ok, ko, partial]

        severity:
          type: string
          enum: [low, medium, high]

        description:
          type: string

        performedAt:
          type: string
          format: date-time

        durationMinutes:
          type: number
          description: Durata intervento in minuti

        calendarEventId:
          type: string
          nullable: true

        createdAt:
          type: string
          format: date-time

        updatedAt:
          type: string
          format: date-time
    Request:
      type: object
      properties:
        _id:
          type: string
        requestType:
          type: string
          enum: [ASSIGN_ROLE, ASSIGN_BUILDING]
        userId:
          type: string
        role:
          type: string
          nullable: true
        buildingId:
          type: string
          nullable: true
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        createdAt:
          type: string
          format: date-time
...